pyproject.toml
[tool.poetry]
name = "conbuddy"
version = "0.1.1a0"
description = "The conbuddy package makes it easy to manage environment specific configuration values."
authors = ["SamEdwardes <edwardes.s@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/samedwardes/conbuddy"
documentation = "https://github.com/samedwardes/conbuddy"
homepage = "https://github.com/samedwardes/conbuddy"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
rich = "^12.6.0"
pydantic = "^1.10.2"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
